1. Refactoring: Connections to Microblogging systems (Twitter, identi.ca) separated from HTTP connections (basic or OAuth) 2. Fixed Issue 36

+++ src/org/andstatus/app/data/MyPreferences.java
@@ -444,7 +444,7 @@ public class MyPreferences {
             themeName.append("Light.");
         }
         themeName.append(name);
-        int themeId = (int) context.getResources().getIdentifier(themeName.toString(), "style",
+        int themeId = context.getResources().getIdentifier(themeName.toString(), "style",
                 "org.andstatus.app");
         if (MyLog.isLoggable(TAG, Log.VERBOSE)) {
             Log.v(TAG, "loadTheme; theme=\"" + themeName.toString() + "\"; id=" + Integer.toHexString(themeId));
